Changed the security features Delay and Lockout In the first versions of mosparo, both security features were active simultaneously. If only the lockout was active, and a user refreshed a page too often within a short time frame, the user got a lockout.
With this change, refreshing a form only count for the Delay security feature, while validating the form (clicking the mosparo checkbox) counts only for the lockout feature.

With this change, a user can refresh a form as often as needed, and only the delay will be appended if enabled. The Lockout feature will only block when the user tries to validate the form.

See #15