Add an automatic update check and info in the header Usually, you have to click the "Check for update" button in the update management interface to see if a new version is available.
If you need to check it manually, you may not get the latest version of mosparo because you need to remember it. This can be a problem, especially if we have to release a security release.

With this change, we've added an automated check. The automated check will be executed once a user with the administrator role has logged in. The information will be cached for one day (24 hours).

In the header of mosparo - next to the "Administration" navigation item - you'll see the button that an update is available. If you click on this button, you will visit the update management interface, where you can start the update.

If you don't want this automatic check, disable it by setting the environment variable "MOSPARO_AUTOMATIC_UPDATE_CHECK_ENABLED" to 0.

See #17