Added the testing of the internet connection mosparo has to check if an internet connection is possible before trying to connect to the internet. Otherwise, a critical error is created.
Some factors do prevent a PHP script from connecting to the internet. First, the PHP setting `allow_url_fopen` can be set to `Off`. In that case, a connection with native PHP functions is not allowed.

The other option is to use the `curl` extension. This extension works even with `allow_url_fopen` set to `Off`. But some security recommendations recommend adding the functions `curl_exec` and `curl_multi_exec` to the PHP setting `disable_functions`, which prevents the execution of these functions.

mosparo can handle all cases and show an error message if the web host's settings do not allow downloads.

See #23