Added the HTTP security headers To enhance the security of mosparo, we've added modern security headers to prevent XSS and other attacks.
For this, we've added the nelmio/security-bundle. It automatically adds the best-suited CSP headers. Additionally, it adds the required hashes for the inline JavaScript blocks.

The bundle also adds the X-XSS-Protection and the X-Frame-Options headers.

All issues regarding inline styles and `data:` URIs were fixed and solved with different approaches. We had to add an additional controller action for the logo in the frontend box to set the correct CORS header.

We must use the `unsafe-inline` flag for styles and the `data:` flag for images on the design settings page. This is because of the color picker, which we must replace with a different one. Since this is a more significant change and we will refactor the design settings in v0.4 anyways, we accepted these unsafe options here for now.