Fixed the fallback session handler for the setup On some systems, the session save path is not writable by the PHP process. Using the `NativeFileSessionHandler` will throw an error on such a system since the session could not be saved correctly.
When we added the `DatabaseSessionHandler`, we misunderstood the Symfony documentation about the default session handler and initialized a `NativeFileSessionHandler`, which tries to save the session to a file. But the default PHP session handler was used before mosparo version v0.3.9.

With this fix, we correctly use the default PHP session handler for the session in the setup.