Refactored the update system to allow upgrades With a new major version, the requirements for the hosting can change. All requirements must be met before the update system can allow such an upgrade.
We've refactored the update repository, which provides the update information. Every major version will have a separate index in the repository. With that, it is possible to maintain multiple major versions simultaneously (for example, maintaining a stable one while testing a new one).

The repository provides the requirements and will be checked by the update system before upgrading to a new version.