Added an option to override the box messages (#72) We've added an option, `customMessages`, to the frontend box options with which it is possible to override the messages in the box.
The option accepts a JavaScript object. The key is the locale's name (for example, `en` or `en_US`), while the value is a JavaScript object. In this object, the key is the name of the message, while the value is the translated message.

The used key in the language object is the same as in line 44 in mosparo-frontend.js.

With this new option, it is also possible to add translations for languages not yet included in mosparo (for example, `de_AT` or `en_AU`).

The frontend uses the `navigator.languages` property to determine the best translation. If a key is not set in the language object or the translated string is empty, the one received from mosparo will be used.

The locale key is written with an underscore (`_`), not a dash (`-`).