Refactored the cleanup logic The cleanup logic could only handle small amounts of submit tokens and submissions. The problem was that we loaded the whole object for submissions and submit tokens. If you have a lot of submissions and submit tokens, it will take too much memory, and will not be able to finish the process.
With the new cleanup logic, we use queries to optimize the process. Additionally, the cleanup within the frontend requests will be limited (delete max. 25k submissions and take max. 1.5 seconds to execute the cleanup).

The cleanup logic will be executed more often now (every 6 hours). If the cleanup is not finished within the limits defined above (25k submissions or 1.5 seconds), it will be executed every 10 minutes until the database is cleaned again.