Added the better and limitless statistics (#151). The statistics for the dashboard chart or the statistics API were calculated by querying all submissions from the database. This was optimized in one of the previous versions but was not done as best, especially since the submissions get deleted after 14 days.
With these refactored statistics, the numbers get stored forever in mosparo. The duration of how long the statistics are stored is configurable per project. The dashboard lets the user decide which range the chart should display. In the API, every date in the past can be selected as a starting point.