Add compatibility with prefixed installations (#216) With this commit, we've added compatibility with the X-Forwarded-Prefix header. With this header, you can set up mosparo with a reverse proxy on a subpath of a domain (example.com/mosparo).
While we recommend using a domain or subdomain and hosting mosparo in the root of that (sub) domain, it's still possible to set up mosparo in a subpath with a reverse proxy.

For this, we've added the required fixes to accept the X-Forwarded-Prefix header, which defines the subpath under which mosparo is served.