Added a fix for database PHP extensions Since we're allowing both MySQL and PostgreSQL, we must add special handling for the required PHP extensions for the upgrade.
The upgrade requirements specify the pdo_pgsql as optional since using a different syntax would not be backward compatible.

Since the pdo_pgsql is not optional if you use PostgreSQL, we have to merge the upgrade requirements and combine the `pdo_mysql` with the `pdo_pgsql` extension to require at least one of these two extensions.